pyproject.toml
[project]
name = "flexpictcut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pyinstaller>=6.17.0",
    "tkinterdnd2>=0.4.3",
]
